LayoutTests/crypto/subtle/aes-export-key.html

<!DOCTYPE html>
<html>
<head>
<script src="../../resources/js-test-pre.js"></script>
<script src="resources/common.js"></script>
</head>
<body>
<p id="description"></p>
<div id="console"></div>

<script>
description("Test exporting an AES key.");

jsTestIsAsync = true;

if (!window.subtle)
    window.crypto.subtle = window.crypto.webkitSubtle;

var extractable = true;

var jwkKey = {
    "kty": "oct",
    "k": "jnOw99oOZFLIEPMrgJB55WL46tJSLGt7"
};

var jwkKeyAsArrayBuffer = asciiToArrayBuffer(JSON.stringify(jwkKey));

debug("Importing a JWK key...");
crypto.subtle.importKey("jwk", jwkKeyAsArrayBuffer, "AES-CBC", extractable, []).then(function(result) {
    key = result;

    debug("Exporting the key as raw data...");
    return crypto.subtle.exportKey("raw", key);
}).then(function(result) {
    exportedData = result;
    shouldBe("byteArrayToHexString(new Uint8Array(exportedData))", "'[8e 73 b0 f7 da 0e 64 52 c8 10 f3 2b 80 90 79 e5 62 f8 ea d2 52 2c 6b 7b]'");

    finishJSTest();
});
</script>

<script src="../../resources/js-test-post.js"></script>
</body>
</html>
